% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaalphaLREP.R
\name{sigmaalphaLREP}
\alias{sigmaalphaLREP}
\title{The program will take your data as input and return three numbers as output: estimate of s, estimate of α, and value of the statistic L.}
\usage{
sigmaalphaLREP(x, tolerance)
}
\arguments{
\item{x}{Import Raw data}

\item{tolerance}{Tolerance quantity}
}
\value{
Output estimate of s, estimate of α, and value of the statistic L.
}
\description{
The program will take your data as input and return three numbers as output: estimate of s, estimate of α, and value of the statistic L.
}
